requirements.txt
sentencepiece==0.2.0
gensim==4.3.2
numpy
transformers>=4.0.0
scipy>=1.12.0,<1.13.0
sklearn-crfsuite==0.3.6
tqdm==4.66.3
ftfy==6.2.0
emoji==1.7.0
requests
nltk
